+++ src/zope/site/next.py
@@ -1,2 +1,3 @@
-# BBB
-from zope.component import getNextUtility, queryNextUtility
+# BBB. Remove in Version 5.0
+from zope.deprecation import moved
+moved('zope.component', 'Version 5.0')